refactor(types): separate test and production type definitions
- Remove WordPress global types from src/types.ts
- Create tests/global.d.ts for test-specific type definitions
- Clean up tests/setup.ts to remove duplicate type interfaces
- Replace any types with type-safe MockedFunction definitions
- Improve type safety by eliminating any type usage
